fix(core) 将Emitter抽取到单独文件，修复在esbuild编译时循环引用无法正确取值的问题
