+++ packages/core/src/common/utils/cancellation.ts
@@ -1,4 +1,5 @@
-import { Event, Emitter } from './event';
+import { Event } from './event';
+import { Emitter } from './emitter';
 
 export interface CancellationToken {
     readonly isCancellationRequested: boolean;
